integrating the rabbitmq into my project

+++ graphql/resolver/book.js
@@ -1,50 +1,57 @@
-const Book = require("../../models/book");
-const { getAuthor, getGenres } = require("./helpers");
+const Book = require('../../models/book');
+const Author=require('../../models/author')
+const { publishNewBookNotification } = require('../../rabbitmq/publisher');
+const { getAuthor, getGenres, getAllAuthorEmails } = require('./helpers');
 
 module.exports = {
-  books: async (args,req) => {
-    if(!req.isAuth){
-        throw new Error('unauthenticated')
-      }
+  books: async () => {
     try {
       const books = await Book.find();
 
       if (!books || books.length === 0) {
         return [];
       }
-      return books.map((book) => {
-        return {
-          ...book._doc,
-          _id: book.id,
-          publicationDate: new Date(book._doc.publicationDate).toISOString(),
-          author: getAuthor.bind(this, book.author),
-          genres: getGenres.bind(this, book.genres),
-        };
-      });
+
+      return books.map((book) => ({
+        ...book._doc,
+        _id: book.id,
+        publicationDate: new Date(book._doc.publicationDate).toISOString(),
+        author: getAuthor.bind(this, book.author),
+        genres: getGenres.bind(this, book.genres),
+      }));
     } catch (err) {
       throw err;
     }
   },
 
-  createBook: async (args,req) => {
-    if(!req.isAuth){
-        throw new Error('unauthenticated')
-      }
-    console.log(req.authorId,"heee");
+  createBook: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     try {
       const book = new Book({
         title: args.bookInput.title,
         description: args.bookInput.description,
         publicationDate: new Date(args.bookInput.publicationDate),
-        author: req.authorId,
+        author: req.authorId,//"65eb7da8d1d90288423e76a4"
         genres: args.bookInput.genreIds,
       });
 
       const result = await book.save();
 
       if (!result) {
-        throw new Error("Book creation failed.");
+        throw new Error('Book creation failed.');
       }
+      
+      
+      const authorEmails=await getAllAuthorEmails()
+
+      const bookData = {
+        title: result.title,
+        authorEmails,
+      };
+      publishNewBookNotification(bookData);
+
 
       return {
         ...result._doc,
@@ -58,21 +65,27 @@ module.exports = {
     }
   },
 
-  updateBook: async ({ id, bookInput },req) => {
-    if(!req.isAuth){
-        throw new Error('unauthenticated')
-      }
+  updateBook: async ({ id, bookInput }, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     try {
       const book = await Book.findById(id);
 
       if (!book) {
-        throw new Error("Book not found.");
+        throw new Error('Book not found.');
+      }
+      // console.log("id",book.author.toString)
+
+      if (book.author.toString() !== req.authorId) {
+        throw new Error('You are not authorized to update this book.');
       }
 
+      
       book.title = bookInput.title;
       book.description = bookInput.description;
       book.publicationDate = new Date(bookInput.publicationDate);
-      book.author = req.authorId;
+      book.author = req.authorId; 
       book.genres = bookInput.genreIds;
 
       const updatedBook = await book.save();
@@ -80,9 +93,7 @@ module.exports = {
       return {
         ...updatedBook._doc,
         _id: updatedBook.id,
-        publicationDate: new Date(
-          updatedBook._doc.publicationDate
-        ).toISOString(),
+        publicationDate: new Date(updatedBook._doc.publicationDate).toISOString(),
         author: getAuthor.bind(this, updatedBook._doc.author),
         genres: getGenres.bind(this, updatedBook._doc.genres),
       };
@@ -91,23 +102,37 @@ module.exports = {
     }
   },
 
-  deleteBook: async ({ id },req) => {
-    if(!req.isAuth){
-        throw new Error('unauthenticated')
-      }
+  deleteBook: async ({ id }, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     try {
+      const book = await Book.findById(id);
+
+      if (!book) {
+        throw new Error('Book not found');
+      }
+
+      if (book.author.toString() !== req.authorId) {
+        throw new Error('You are not authorized to delete this book.');
+      }
+
       const deletedBook = await Book.findByIdAndDelete(id);
 
       if (!deletedBook) {
-        throw new Error("Book not found");
+        throw new Error('Book not found');
       }
 
       return {
         ...deletedBook._doc,
         _id: deletedBook.id,
+        publicationDate: new Date(deletedBook._doc.publicationDate).toISOString(),
+        author: getAuthor.bind(this, deletedBook._doc.author),
+        genres: getGenres.bind(this, deletedBook._doc.genres),
       };
     } catch (err) {
       throw err;
     }
   },
+  
 };
